+++ packages/adminui-vue/components/controls/UDropdown/UDropdown.vue
@@ -208,10 +208,12 @@ export default {
       this.observer = new MutationObserver(() => this.checkAndUpdatePopupPosition())
 
       setTimeout(() => {
-        this.observer.observe(this.$refs.dropdown, {
-          childList: true,
-          subtree: true
-        })
+        if (this.$refs.dropdown) { // for slow clients this.$refs.dropdown can be undefined on this stage
+          this.observer.observe(this.$refs.dropdown, {
+            childList: true,
+            subtree: true
+          })
+        }
       }, 0)
     },