+++ apps/autotest/models/TST/public/forms/tst_service-crypto-fm.vue
@@ -242,7 +242,7 @@ module.exports.default = {
       } finally {
         // me.unmask()
       }
-    },
+    }
   }
 }
 </script>