+++ packages/adminui-reg/index.html.js
@@ -13,9 +13,11 @@ const MODULES_ROOT = path.join(process.configPath, 'node_modules')
 
 /**
  * Resolve a path to module entry point by module name
+ *
  * @param {string} moduleName
  * @param {Array<{moduleName: string, entryPoint: string}>} modulesMap
- * @return {undefined|string}
+ * @param modelsConfig
+ * @returns {undefined|string}
  */
 function resolveModuleEntryPoint (moduleName, modulesMap, modelsConfig) {
   if (modulesMap.find(m => m.moduleName === moduleName)) return // already added
@@ -24,11 +26,11 @@ function resolveModuleEntryPoint (moduleName, modulesMap, modelsConfig) {
   if (!fs.existsSync(resolvedPath)) { // try js file
     resolvedPath = resolvedPath + '.js'
   }
-  let stat = fs.statSync(resolvedPath)
+  const stat = fs.statSync(resolvedPath)
   if (stat.isDirectory()) {
-    let pkgName = path.join(resolvedPath, 'package.json')
+    const pkgName = path.join(resolvedPath, 'package.json')
     if (fs.existsSync(pkgName)) {
-      let pkgMain = JSON.parse(fs.readFileSync(pkgName, 'utf8')).main || './index.js'
+      const pkgMain = JSON.parse(fs.readFileSync(pkgName, 'utf8')).main || './index.js'
       resolvedPath = path.join(resolvedPath, pkgMain)
     } else {
       resolvedPath = path.join(resolvedPath, 'index.js')
@@ -39,6 +41,12 @@ function resolveModuleEntryPoint (moduleName, modulesMap, modelsConfig) {
   }
 }
 
+/**
+ * @param cspAllow
+ * @param cspAllowP
+ * @param name
+ * @param init
+ */
 function initCspAllow (cspAllow, cspAllowP, name, init) {
   if (!cspAllow[name]) {
     cspAllow[name] = {}
@@ -53,19 +61,22 @@ function initCspAllow (cspAllow, cspAllowP, name, init) {
  *
  * @param {THTTPRequest} req
  * @param {THTTPResponse} resp
- * @param {String} indexName
+ * @param {string} indexName
  * @param {boolean} [addCSP=true] Add a CSP header
  */
 function generateIndexPage (req, resp, indexName, addCSP = true) {
+  /**
+   * @param modelCode
+   */
   function modelVer (modelCode) {
-    let m = App.domainInfo.models[modelCode]
+    const m = App.domainInfo.models[modelCode]
     return m ? `?ver=${m.version}` : '?ver=0'
   }
 
-  let uiSettings = App.serverConfig.uiSettings || {}
+  const uiSettings = App.serverConfig.uiSettings || {}
   let cspAllow = uiSettings.cspAllow || {}
-  let cspAllowP = {}
-  let compiledIndexKey = GC_KEYS.COMPILED_INDEX_ + indexName
+  const cspAllowP = {}
+  const compiledIndexKey = GC_KEYS.COMPILED_INDEX_ + indexName
   let compiledIndex = App.globalCacheGet(compiledIndexKey)
   if (!compiledIndex) {
     if (!uiSettings.adminUI) {
@@ -80,16 +91,26 @@ function generateIndexPage (req, resp, indexName, addCSP = true) {
         uiSettings.adminUI.loginURL = '/models/adminui-pub/login-stub.html'
       }
     }
-    let adminUIPath = path.dirname(require.resolve('@unitybase/adminui-pub'))
-    let indexTpl = fs.readFileSync(path.join(adminUIPath, indexName), 'utf8')
+    const adminUIPath = path.dirname(require.resolve('@unitybase/adminui-pub'))
+    const indexTpl = fs.readFileSync(path.join(adminUIPath, indexName), 'utf8')
 
-    let cspNonce = nsha256('' + Date.now()).substr(0, 8)
+    const cspNonce = nsha256('' + Date.now()).substr(0, 8)
     App.globalCachePut(GC_KEYS.COMPILED_INDEX_NONCE, cspNonce)
+    let customThemeCSS, customThemeJS
+    debugger
+    if (uiSettings.adminUI.customTheme) {
+      const allThemes = ubm_desktop.getUIThemes()
+      const ct = allThemes.find(t => t.name === uiSettings.adminUI.customTheme)
+      if (ct) {
+        customThemeCSS = ct.css
+        customThemeJS = ct.js
+      }
+    }
     // create view for mustache
     // noinspection JSUnusedGlobalSymbols
-    let view = {
+    const view = {
       uiSettings: uiSettings,
-      modulesMap: [ // modules with entry point path differ from specified in package.json main section. In futur ewill be used for ESM module import configuration
+      modulesMap: [ // modules with entry point path differ from specified in package.json main section. In future will be used for ESM module import configuration
         { moduleName: 'css', entryPoint: 'systemjs-plugin-css/css.js' },
         { moduleName: 'vue', entryPoint: 'vue/dist/vue.common.dev.js' }, // should be the same as in adminui-vue webpack config
         { moduleName: 'element-ui', entryPoint: 'element-ui/lib/index.js' },
@@ -103,13 +124,15 @@ function generateIndexPage (req, resp, indexName, addCSP = true) {
       UB_API_PATH: App.serverConfig.httpServer.path || '/',
       modelVer: function () {
         return modelVer
-      }
+      },
+      customThemeCSS,
+      customThemeJS
     }
 
     // add admin-ui and models what require initialisation
-    let modelsConfig = App.serverConfig.application.domain.models
+    const modelsConfig = App.serverConfig.application.domain.models
     const ADMINUI_MODEL = '@unitybase/adminui-pub'
-    let modelCfg = modelsConfig.find(m => m.moduleName === ADMINUI_MODEL)
+    const modelCfg = modelsConfig.find(m => m.moduleName === ADMINUI_MODEL)
     view.adminUIModel = modelCfg.browser
     App.domainInfo.orderedModels.forEach((model) => {
       // fill model versions
@@ -120,13 +143,13 @@ function generateIndexPage (req, resp, indexName, addCSP = true) {
         })
       }
       if (model.moduleName !== ADMINUI_MODEL) {
-        let modelCfg = modelsConfig.find(m => m.name === model.name)
+        const modelCfg = modelsConfig.find(m => m.name === model.name)
         if (modelCfg.browser) {
           view.modelInitialization.push(modelCfg.browser)
         } else if (model.needInit) { // ub 5.0.5 compatibility
           console.warn(`Compatibility warning! Model ${model.name} has browser initialization script,
 but "browser" section in package.json is not defined. Will fallback to "browser": "./public/initModel.js"`)
-          let pathStart = /node_modules/.test(model.path)
+          const pathStart = /node_modules/.test(model.path)
             ? model.moduleName + '/public'
             : path.join('/clientRequire', model.path).replace(/\\/g, '/')
           view.modelInitialization.push({
@@ -177,7 +200,7 @@ but "browser" section in package.json is not defined. Will fallback to "browser"
     // cache forever - do not cache index*.html
     let cspHeader = ''
     if (addCSP) {
-      let cspNonce = App.globalCacheGet(GC_KEYS.COMPILED_INDEX_NONCE)
+      const cspNonce = App.globalCacheGet(GC_KEYS.COMPILED_INDEX_NONCE)
       // let wsSrc = 'ws' + App.externalURL.slice(4)
       // let uiSettings = App.serverConfig.uiSettings
       // let onlyOfficeServer = (uiSettings.adminUI.onlyOffice && uiSettings.adminUI.onlyOffice.serverIP) || ''
@@ -195,8 +218,8 @@ but "browser" section in package.json is not defined. Will fallback to "browser"
       initCspAllow(cspAllow, cspAllowP, 'fontSrc')
       initCspAllow(cspAllow, cspAllowP, 'baseUri', 'resource:')
 
-      let wsNotifier = App.serverConfig.uiSettings.adminUI.amqpNotificationUrl || ''
-      let cspHeaders =
+      const wsNotifier = App.serverConfig.uiSettings.adminUI.amqpNotificationUrl || ''
+      const cspHeaders =
         `default-src 'self' ${cspAllowP.defaultSrc}; ` +
         `connect-src 'self' ${wsNotifier} ${cspAllowP.connectSrc} blob:; ` + // we need blob: for UBDocument (ER diagrams, org chart etc.)
         // 'unsafe-inline' is removed in flavor of 'nonce-...'
@@ -208,7 +231,7 @@ but "browser" section in package.json is not defined. Will fallback to "browser"
         `font-src 'self' data: ${cspAllowP.fontSrc}; ` +
         `frame-src 'self' ${cspAllowP.frameSrc} blob:; ` + // blob src required for chrome PDF viewer. Self - for JS PDF viewer
         `img-src 'self' https://unitybase.info ${cspAllowP.imgSrc} data: blob:; ` + // blob: is for pictures inside tinyMCE
-        `plugin-types application/pdf`
+        'plugin-types application/pdf'
       cspHeader = '\r\nContent-Security-Policy: ' + cspHeaders
       console.debug(cspHeaders)
     }