+++ packages/adminui-vue/CHANGELOG.md
@@ -6,10 +6,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - sidebar (instance of USidebar component) is exported as `SidebarInstance`. Example below collapse sidebar programmatically:
+```javascript
+const { SidebarInstance } = require('@unitybase/adminui-vue')
+SidebarInstance.isCollapsed = true
+``` 
+
+ - `URadio`: `id` and `label` properties names can be customized 
 
 ### Changed
  - `UDialog`: width of the dialog box is extended up to `540px` in order to keep
-   all dialog buttons on the same level. Some buttons may have too long captions
+   all dialog buttons on the same level. Some buttons captions are too long 
    (as for the `de` locale).
  - `UDropdown`: fixed `Uncaught TypeError: Failed to execute 'observe' on 'MutationObserver'` for slow client machine